move channel getter to querier file
