+++ cw-orch/tests/querier.rs
@@ -3,9 +3,12 @@ mod common;
 #[cfg(feature = "node-tests")]
 mod queriers {
     use crate::common;
+    use cw_orch::{daemon::channel::GrpcChannel, prelude::networks};
+    use ibc_chain_registry::chain::Grpc;
+    use ibc_relayer_types::core::ics24_host::identifier::ChainId;
+    use speculoos::{asserting, result::ResultAssertions};
     use std::{str::FromStr, sync::Arc};
 
-    use common::channel::build_channel;
     use cw_orch::{
         daemon::{
             error::DaemonError,
@@ -13,8 +16,6 @@ mod queriers {
         },
         prelude::{queriers::StakingBondStatus, *},
     };
-
-    use speculoos::prelude::*;
     use tokio::runtime::Runtime;
 
     use cosmrs::{
@@ -22,6 +23,25 @@ mod queriers {
         tx::{self, Msg},
         AccountId, Denom,
     };
+
+    pub async fn build_channel() -> tonic::transport::Channel {
+        let network = networks::LOCAL_JUNO;
+
+        let grpcs: Vec<Grpc> = vec![Grpc {
+            address: network.grpc_urls[0].into(),
+            provider: None,
+        }];
+
+        let chain: ChainId = ChainId::new(network.chain_id.to_owned(), 1);
+
+        let channel = GrpcChannel::connect(&grpcs, &chain).await;
+
+        asserting!("channel connection is succesful")
+            .that(&channel)
+            .is_ok();
+
+        channel.unwrap()
+    }
 
     /*
         Querier - Ibc
